Move s3 bucket code to its own file
